separate into files and rename workflow
