+++ src/models/searchservice/response/directanswer/FieldValueDirectAnswer.ts
@@ -1,5 +1,13 @@
-import { AndroidAppUrlDirectAnswer, DecimalDirectAnswer, FacebookUrlDirectAnswer, InstagramHandleDirectAnswer, IosAppUrlDirectAnswer, RichTextDirectAnswer, TextValueDirectAnswer, TwitterHandleDirectAnswer, UrlDirectAnswer } from './StringValueDirectAnswer';
-import { UnknownFieldValueDirectAnswer } from './UnknownFieldValueDirectAnswer';
+import { UnknownFieldValueDirectAnswer } from './fieldvaluedirectanswers/UnknownFieldValueDirectAnswer';
+import { DecimalDirectAnswer } from './fieldvaluedirectanswers/DecimalDirectAnswer';
+import { InstagramHandleDirectAnswer } from './fieldvaluedirectanswers/InstagramHandleDirectAnswer';
+import { IosAppUrlDirectAnswer } from './fieldvaluedirectanswers/IosAppUrlDirectAnswer';
+import { AndroidAppUrlDirectAnswer } from './fieldvaluedirectanswers/AndroidAppUrlDirectAnswer';
+import { TwitterHandleDirectAnswer } from './fieldvaluedirectanswers/TwitterHandleDirectAnswer';
+import { FacebookUrlDirectAnswer } from './fieldvaluedirectanswers/FacebookUrlDirectAnswer';
+import { RichTextDirectAnswer } from './fieldvaluedirectanswers/RichTextDirectAnswer';
+import { UrlDirectAnswer } from './fieldvaluedirectanswers/UrlDirectAnswer';
+import { TextValueDirectAnswer } from './fieldvaluedirectanswers/TextValueDirectAnswer';
 
 /**
  * Possible built-in and custom {@link BaseFieldValueDirectAnswer} interfaces.
