+++ src/models/searchservice/response/directanswer/fieldvaluedirectanswers/TextValueDirectAnswer.ts
@@ -1,0 +1,12 @@
+import { EnumOrLiteral } from '../../../../utils/EnumOrLiteral';
+import { BaseFieldValueDirectAnswer } from '../BaseFieldValueDirectAnswer';
+import { BuiltInFieldType } from '../BuiltInFieldType';
+
+/**
+ * A direct answer whose source is a string or string list field in the knowledge graph.
+ *
+ * @public
+ */
+export interface TextValueDirectAnswer extends BaseFieldValueDirectAnswer<string | string[]> {
+  fieldType: EnumOrLiteral<BuiltInFieldType.SingleLineText | BuiltInFieldType.MultiLineText>
+}
