+++ src/models/searchservice/response/directanswer/fieldvaluedirectanswers/TwitterHandleDirectAnswer.ts
@@ -1,0 +1,12 @@
+import { EnumOrLiteral } from '../../../../utils/EnumOrLiteral';
+import { BaseFieldValueDirectAnswer } from '../BaseFieldValueDirectAnswer';
+import { BuiltInFieldType } from '../BuiltInFieldType';
+
+/**
+ * A direct answer for an twitter handle field.
+ *
+ * @public
+ */
+export interface TwitterHandleDirectAnswer extends BaseFieldValueDirectAnswer<string> {
+  fieldType: EnumOrLiteral<BuiltInFieldType.TwitterHandle>
+}
