+++ src/models/searchservice/response/directanswer/fieldvaluedirectanswers/index.ts
@@ -1,0 +1,10 @@
+export { UnknownFieldValueDirectAnswer } from './UnknownFieldValueDirectAnswer';
+export { DecimalDirectAnswer } from './DecimalDirectAnswer';
+export { InstagramHandleDirectAnswer } from './InstagramHandleDirectAnswer';
+export { IosAppUrlDirectAnswer } from './IosAppUrlDirectAnswer';
+export { AndroidAppUrlDirectAnswer } from './AndroidAppUrlDirectAnswer';
+export { TwitterHandleDirectAnswer } from './TwitterHandleDirectAnswer';
+export { FacebookUrlDirectAnswer } from './FacebookUrlDirectAnswer';
+export { RichTextDirectAnswer } from './RichTextDirectAnswer';
+export { UrlDirectAnswer } from './UrlDirectAnswer';
+export { TextValueDirectAnswer } from './TextValueDirectAnswer';
